@@ -30,7 +30,7 @@
     "test": "mocha 'test/*_test.js'"
   },
   "engines": {
-    "node": ">=14.0.0",
+    "node": ">=18.0.0",
     "npm": ">=7.0.0"
   }
 }
