@@ -40,12 +40,12 @@
     "test": "c8 mocha -ui tdd --require ./test/index.js 'test/*_test.js'"
   },
   "engines": {
-    "node": ">=14.0.0",
+    "node": ">=18.0.0",
     "npm": ">=7.0.0"
   },
   "devDependencies": {
     "@fluent/bundle": "^0.18.0",
-    "jsdom": "^21.1.0"
+    "jsdom": "^24.0.0"
   },
   "dependencies": {
     "cached-iterable": "^0.3"
