@@ -23,31 +23,30 @@
     "test": "npm run test:common && npm test -w fluent-dom && npm test -w fluent-react"
   },
   "engines": {
-    "node": ">=14.18.0",
+    "node": ">=18.0.0",
     "npm": ">=7.0.0"
   },
   "prettier": {
-    "arrowParens": "avoid"
+    "arrowParens": "avoid",
+    "trailingComma": "es5"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.21.0",
-    "@typescript-eslint/parser": "^5.21.0",
-    "c8": "^7.7.3",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "c8": "^9.1.0",
     "colors": "^1.3.3",
     "commander": "^2.20",
     "eslint": "^8.14.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-mocha": "^10.1.0",
-    "esm": "^3.2.22",
     "fuzzer": "^0.2.1",
-    "gh-pages": "^5.0.0",
-    "intl-pluralrules": "^1.0.0",
+    "gh-pages": "^6.1.1",
     "mocha": "^10.2.0",
-    "prettier": "^2.6.2",
+    "prettier": "^3.2.5",
     "prettyjson": "^1.2.1",
-    "rollup": "^3.15.0",
-    "sinon": "^15.0.1",
-    "typedoc": "^0.23.25",
-    "typescript": "^4.4.3"
+    "rollup": "^4.12.0",
+    "sinon": "^17.0.1",
+    "typedoc": "^0.25.9",
+    "typescript": "^5.3.3"
   }
 }
